data-fetcher/news/requirements.txt
feedparser==6.0.11
PyMySQL==1.1.0
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
beautifulsoup4==4.12.3
lxml==5.1.0

